requirements.txt
# Web Framework
fastapi
uvicorn
pydantic
pydantic-settings

# Database
sqlalchemy
psycopg2-binary
chromadb

# AI & LLM
langchain
langchain-openai
langgraph
openai
tiktoken
langsmith

# Frontend & UI
streamlit
requests

# Utilities
python-dotenv